Inspired by #1242, updates the aggregation framework to run all pipelines multi-threaded, even ones that aren't facet-able.

```py
import eta.core.utils as etau
import fiftyone as fo
from fiftyone import ViewField as F

dataset = fo.load_dataset("bdd100k-validation")

with etau.Timer():
    ids = dataset.values("id")

with etau.Timer():
    filepaths = dataset.values("filepath", _big_result=False)

with etau.Timer():
    ids, filepaths = dataset.values(["id", "filepath"])

with etau.Timer():
    ids, filepaths = dataset.values(["id", "filepath"], _big_result=False)

with etau.Timer():
    results = dataset.aggregate(
        [
            fo.CountValues("detections.detections.label"),
            fo.Distinct("detections.detections.label"),
            fo.Distinct("scene.label"),
            fo.Values("id"),
            fo.Values("filepath", _big_result=False),
            fo.Values("detections.detections.label"),
            fo.CountValues("polylines.polylines.filled"),
            fo.Bounds(F("polylines.polylines.points").length()),
            fo.Values(F("polylines.polylines.points").length()),
        ]
    )

```

Typical results:

```
### facet-optimization2
Time elapsed: 62.3ms
Time elapsed: 18.1ms
Time elapsed: 53.0ms
Time elapsed: 101.4ms
Time elapsed: 842.2ms

### facet-optimization
Time elapsed: 55.9ms
Time elapsed: 21.0ms
Time elapsed: 62.5ms
Time elapsed: 102.2ms
Time elapsed: 1.6 seconds

### develop
Time elapsed: 46.8ms
Time elapsed: 24.7ms
Time elapsed: 57.6ms
Time elapsed: 31.3ms
Time elapsed: 2.5 seconds
```