@brimoor 2021-09-03
@benjaminpkane FYI, I discovered that it *is* in fact better to put the `list()` inside the thread pool, like you had it. Now this PR is a strict improvement 🥳